refactor(Slack): SlackBoardInfo 패키지 변경
